add missing changes of last commit
